src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
attrs==19.3.0
beautifulsoup4==4.9.0     # via bs4
blinker==1.4              # via sentry-sdk
bs4==0.0.1
certifi==2020.4.5.1       # via requests, sentry-sdk
chardet==3.0.4            # via requests
click==7.1.2
flask-cors==3.0.8
flask==1.1.2
idna==2.9                 # via requests
importlib-metadata==1.6.1  # via pluggy, pytest
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
markupsafe==1.1.1         # via jinja2
more-itertools==8.4.0     # via pytest
packaging==20.4           # via pytest
pluggy==0.13.1            # via pytest
py==1.8.2                 # via pytest
pyparsing==2.4.7          # via packaging
pytest==5.4.3
python-trello==0.0.4
redis==3.5.3
requests==2.23.0
rq==1.4.2
selenium==3.141.0
sentry-sdk[flask]==0.15.1
six==1.15.0               # via flask-cors, packaging
soupsieve==2.0            # via beautifulsoup4
urllib3==1.25.9           # via requests, selenium, sentry-sdk
wcwidth==0.2.4            # via pytest
werkzeug==1.0.1           # via flask
zipp==3.1.0               # via importlib-metadata
